requirements.txt
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
pillow
drf-yasg
python-decouple
django-cors-headers
redis
celery
django-filter
djangorestframework-pagination
requests
django-password-validators
aiogram
django-telegram-login
black
isort
flake8